challenge-3-new/requirements.txt
# Challenge 3 Requirements
# Refactored Maintenance Scheduler Agent

# Core dependencies
python-dotenv>=1.0.0
asyncio>=3.4.3

# Azure SDK
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-ai-projects>=1.0.0b4

# Agent Framework
agent-framework>=0.1.0
agent-framework-azure-ai>=0.1.0

# Observability (optional but recommended)
azure-monitor-opentelemetry-exporter>=1.0.0b21

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Type checking (development)
mypy>=1.7.0
